requirements.txt
streamlit>=1.33.0
google-generativeai>=0.4.1
transformers>=4.40.0
sentencepiece>=0.1.99
torch>=2.1.2
PyMuPDF>=1.23.7
